-benchmark-cli: Update

+++ actor4j-benchmark-cli/src/main/java/io/actor4j/benchmark/BenchmarkConfig.java
@@ -68,7 +68,7 @@ public class BenchmarkConfig {
 		return this;
 	}
 	
-	public long iterations() {
+	public int iterations() {
 		return iterations ;
 	}
